+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [0.125.0](https://github.com/go-to-k/cdkd/compare/v0.124.0...v0.125.0) (2026-05-22)
+
+
+### Features
+
+* **local-invoke:** accept literal layer ARNs (cross-account / cross-region) — closes [#448](https://github.com/go-to-k/cdkd/issues/448) ([#491](https://github.com/go-to-k/cdkd/issues/491)) ([9c30f0e](https://github.com/go-to-k/cdkd/commit/9c30f0e4905cfc02f102a034c84c5f5dcc3a0561))
+
+# [0.124.0](https://github.com/go-to-k/cdkd/compare/v0.123.0...v0.124.0) (2026-05-22)
+
+
+### Features
+
+* **export:** YAML template support ([#483](https://github.com/go-to-k/cdkd/issues/483)) ([bc69422](https://github.com/go-to-k/cdkd/commit/bc6942273f44b4a807f8ea42058b6140a78352ec))
+
+# [0.123.0](https://github.com/go-to-k/cdkd/compare/v0.122.0...v0.123.0) (2026-05-22)
+
+
+### Features
+
+* **intrinsic:** Fn::GetStackOutput cross-account RoleArn support ([#490](https://github.com/go-to-k/cdkd/issues/490)) ([ffc284f](https://github.com/go-to-k/cdkd/commit/ffc284f876769aa0ac3e205e3f8cfb01466140c1))
+
 # [0.122.0](https://github.com/go-to-k/cdkd/compare/v0.121.0...v0.122.0) (2026-05-22)